+++ lsdpack.cpp
@@ -170,7 +170,10 @@ void record_gbs(int argc, char* argv[]) {
 }
 
 void print_help_and_exit() {
-    fprintf(stderr, "usage: lsdpack [-g] [lsdj.gb lsdj2.gb ...]\n");
+    puts("usage: lsdpack [-g] [-r] [lsdj.gb lsdj2.gb ...]");
+    puts("");
+    puts("-g: .gbs conversion");
+    puts("-r: raw register writes; optimizations disabled");
     exit(1);
 }
 
@@ -178,11 +181,15 @@ int main(int argc, char* argv[]) {
     bool gbs_mode = false;
 
     int c;
-    while ((c = getopt(argc, argv, "g")) != -1) {
+    while ((c = getopt(argc, argv, "gr")) != -1) {
         switch (c) {
             case 'g':
                 puts(".gbs mode enabled");
                 gbs_mode = true;
+                break;
+            case 'r':
+                puts("disabled optimizations");
+                Writer::disable_optimizations();
                 break;
             default:
                 print_help_and_exit();